BlockedHandler: changed the block handler to: first ack the message and then dispatch to the queue.
ClusterHeartbeat: added the namespace to the constructor
and switched from the lowlevel thrading api "start_new_thread"
to the high level "Thread" class.

ListHandler: reformated using autoformatter

NodeRegistration: New class from node registration code
from the TaskQueue class.

QueueHandler: added the virtual_host variable to be able
to correctly use a separate virtual host in rabbitmq for each namespace.
refactored the code to check the a task
in the _on_message method to a separate method _on_message_check_task
refactored the code to check the task result
in _on_task to _on_task_check_task_result
and there also to _on_task_result and _on_none_task_result

TaskHandler: fixed the self.block_list ListHandler redis key
to also use the namespace and refactored using an autoformatter
The namespace is now also reflacted in the Workflow mongodb logging
Refactored the code to exclude sensitive arguments
in the database to a separate class
Mongodb is now used to store the result
of each workfow and task instead of redis
Most static methods are now part of the Task Model
Before each task invocation is now checked,
if the workflow has already been stopped.
Also formatted using an autoformatter.

TaskQueue: refactored the node registration code to a separate class

Models/Task: Refactored some TaskHandler static methods
for checking if a parent task exists/etc. to the Task model

Models/TaskLog: added the worklflow_id
for faster querying from the backend-api

Models/Workflow: reformatted using an autoformatter

Models/Heartbeat: added the namespace to the model

Wrapper/AMQP: reformatted using an autoformatter
Added a try/except for AMQPConnectionError
to correctly intterupt the main thread to correctly shut down the node
in case of an error in the communication with the rabbitmq server.

BytesIOWrapper: Added the workflow_id
to each log entry in the mongodb database